+++ lib/core/test/element_special_values_test.cpp
@@ -31,7 +31,7 @@ TYPED_TEST_SUITE(ElementIsnanTest, ElementSpecialValuesTestTypes);
 
 TEST(ElementIsnanTest, unit) {
   for (const auto &u : {units::dimensionless, units::m, units::meV}) {
-    EXPECT_EQ(element::isnan(u), units::dimensionless);
+    EXPECT_EQ(element::isnan(u), units::none);
   }
 }
 
@@ -57,7 +57,7 @@ TYPED_TEST_SUITE(ElementIsinfTest, ElementSpecialValuesTestTypes);
 
 TEST(ElementIsinfTest, unit) {
   for (const auto &u : {units::dimensionless, units::m, units::meV}) {
-    EXPECT_EQ(element::isinf(u), units::dimensionless);
+    EXPECT_EQ(element::isinf(u), units::none);
   }
 }
 
@@ -83,7 +83,7 @@ TYPED_TEST_SUITE(ElementIsfiniteTest, ElementSpecialValuesTestTypes);
 
 TEST(ElementIsfiniteTest, unit) {
   for (const auto &u : {units::dimensionless, units::m, units::meV}) {
-    EXPECT_EQ(element::isfinite(u), units::dimensionless);
+    EXPECT_EQ(element::isfinite(u), units::none);
   }
 }
 
@@ -110,8 +110,8 @@ TYPED_TEST_SUITE(ElementIssignedinfTest, ElementSpecialValuesTestTypes);
 
 TEST(ElementIssignedinfTest, unit) {
   for (const auto &u : {units::dimensionless, units::m, units::meV}) {
-    EXPECT_EQ(element::isposinf(u), units::dimensionless);
-    EXPECT_EQ(element::isneginf(u), units::dimensionless);
+    EXPECT_EQ(element::isposinf(u), units::none);
+    EXPECT_EQ(element::isneginf(u), units::none);
   }
 }