+++ lib/variable/bin_detail.cpp
@@ -20,7 +20,7 @@ namespace scipp::variable::bin_detail {
 /// 0 if the coord is less than the first edge, nbin-1 if greater or equal last
 /// edge. Assumes both `edges` and `coord` are sorted.
 Variable begin_edge(const Variable &coord, const Variable &edges) {
-  auto indices = makeVariable<scipp::index>(coord.dims());
+  auto indices = makeVariable<scipp::index>(coord.dims(), units::none);
   const auto dim = edges.dims().inner();
   if (indices.dims()[dim] == 0)
     return indices;
@@ -36,7 +36,7 @@ Variable begin_edge(const Variable &coord, const Variable &edges) {
 /// nbin if the coord is greater than the last edge. Assumes both `edges` and
 /// `coord` are sorted.
 Variable end_edge(const Variable &coord, const Variable &edges) {
-  auto indices = makeVariable<scipp::index>(coord.dims());
+  auto indices = makeVariable<scipp::index>(coord.dims(), units::none);
   const auto dim = edges.dims().inner();
   if (indices.dims()[dim] == 0)
     return indices;