+++ src/scipp/plotting/resampling_model.py
@@ -32,7 +32,7 @@ def _resample(array, mode: ResamplingMode, dim, edges):
     # `rebin` that works on non-counts data
     coord = array.coords[dim]
     width = coord[dim, 1:] - coord[dim, :-1]
-    width.unit = units.one
+    width.unit = None if array.unit is None else units.one
     array.data *= width
     unit = array.unit
     array.unit = units.counts
@@ -286,10 +286,6 @@ class ResamplingBinnedModel(ResamplingModel):
 
 def _with_edges(array):
     new_array = array.copy(deep=False)
-    if new_array.unit is None:
-        # Workaround required, even if no resampling done
-        new_array.data = array.data.copy()
-        new_array.unit = ''
     prefix = ''.join(array.dims)
     for dim, var in array.coords.items():
         if dim not in array.dims: